theme not applying on all pages
