+++ client/src/App.js
@@ -6,12 +6,14 @@ import Routes from './components/routing/Routes';
 import NotFound from './components/layout/NotFound';
 import { LOGOUT } from './actions/types';
 import { makeStyles } from '@material-ui/core/styles';
-
+import Tooltip from '@material-ui/core/Tooltip';
+import CreateIcon from '@material-ui/icons/Create';
 // Redux
 import { Provider } from 'react-redux';
 import store from './store';
 import { loadUser } from './actions/auth';
 import setAuthToken from './utils/setAuthToken';
+import ThemeDialog from "./components/themes/Theme";
 const useStyles = makeStyles((theme) => ({
 
   back: {
@@ -19,7 +21,24 @@ const useStyles = makeStyles((theme) => ({
     backgroundPosition: "center",
     height: "auto",
     minHeight: "100vh"
-  }
+  },
+  editTheme: {
+    color: "white",
+    fontSize: "20px",
+    position: "absolute",
+    right: "7%",
+    bottom: "4%",
+    position: 'fixed',
+    cursor: "pointer",
+    padding: "10px",
+    transition: "0.1s all",
+    "&:hover": {
+      background: "white",
+      padding: "10px",
+      color: "black",
+      borderRadius: "50%",
+    }
+  },
 
 
 }));
@@ -28,6 +47,13 @@ const App = () => {
   const classes = useStyles();
   const [displayBackground, setBackground] = React.useState(false);
   const [theme, setTheme] = React.useState(null);
+  const [open, setOpen] = React.useState(false);
+  const handleClickOpen = () => {
+    setOpen(true);
+  }
+  const handleClose = () => {
+    setOpen(false);
+  }
   useEffect(() => {
     // check for token in LS
     if (localStorage.token) {
@@ -50,7 +76,7 @@ const App = () => {
     <Provider store={store}>
       <Router>
         <Fragment>
-          <div className={classes.back} style={{ backgroundImage: displayBackground && window.location.pathname == "/dashboard" ? `url(${theme})` : null, }}>
+          <div className={classes.back} style={{ backgroundImage: displayBackground && window.location.pathname != "/" ? `url(${theme})` : null, }}>
             <Navbar />
 
             <Switch>
@@ -58,6 +84,10 @@ const App = () => {
               <Route component={Routes} />
               <Route component={NotFound} />
             </Switch>
+            <Tooltip title="Change Theme">
+              <CreateIcon onClick={handleClickOpen} className={classes.editTheme} />
+            </Tooltip>
+            <ThemeDialog open={open} handleClose={handleClose} />
           </div>
