+++ client/src/components/tests/QuestionForm.js
@@ -92,10 +92,10 @@ const QuestionForm = ({
   return (
     <div style={{ padding: 16, margin: 'auto', maxWidth: 600 }}>
       <CssBaseline />
-      <Typography variant="h4" align="center" component="h1" gutterBottom>
+      <Typography style={{ color: "white" }} variant="h4" align="center" component="h1" gutterBottom>
         Create Test
       </Typography>
-      <Typography variant="h5" align="center" component="h2" gutterBottom>
+      <Typography style={{ color: "white" }} variant="h5" align="center" component="h2" gutterBottom>
         Questions Form
       </Typography>
       <Paper style={{ padding: 16 }}>
@@ -115,98 +115,98 @@ const QuestionForm = ({
         <Grid container alignItems="flex-start" spacing={2}>
           {questions
             ? questions.map((curr, index) => {
-                return (
-                  <React.Fragment key={index}>
-                    <Grid item xs={12}>
-                      <TextField
-                        required
-                        id="question"
-                        name="question"
-                        value={curr.question}
-                        label={`Question ${index + 1}`}
-                        fullWidth
-                        onChange={(e) => handleChange(e, index)}
-                      />
-                    </Grid>
-                    <Grid item xs={6}>
-                      <TextField
-                        required
-                        id="option1"
-                        name="option1"
-                        label="Option 1"
-                        value={curr.option1}
-                        fullWidth
-                        onChange={(e) => handleChange(e, index)}
-                      />
-                    </Grid>
-                    <Grid item xs={6}>
-                      <TextField
-                        required
-                        id="option2"
-                        name="option2"
-                        label="Option 2"
-                        value={curr.option2}
-                        fullWidth
-                        onChange={(e) => handleChange(e, index)}
-                      />
-                    </Grid>
-                    <Grid item xs={6}>
-                      <TextField
-                        required
-                        id="option3"
-                        name="option3"
-                        label="Option3"
-                        value={curr.option3}
-                        fullWidth
-                        onChange={(e) => handleChange(e, index)}
-                      />
-                    </Grid>
-                    <Grid item xs={6}>
-                      <TextField
-                        required
-                        id="option4"
-                        name="option4"
-                        label="Option4"
-                        value={curr.option4}
-                        fullWidth
-                        onChange={(e) => handleChange(e, index)}
-                      />
-                    </Grid>
+              return (
+                <React.Fragment key={index}>
+                  <Grid item xs={12}>
+                    <TextField
+                      required
+                      id="question"
+                      name="question"
+                      value={curr.question}
+                      label={`Question ${index + 1}`}
+                      fullWidth
+                      onChange={(e) => handleChange(e, index)}
+                    />
+                  </Grid>
+                  <Grid item xs={6}>
+                    <TextField
+                      required
+                      id="option1"
+                      name="option1"
+                      label="Option 1"
+                      value={curr.option1}
+                      fullWidth
+                      onChange={(e) => handleChange(e, index)}
+                    />
+                  </Grid>
+                  <Grid item xs={6}>
+                    <TextField
+                      required
+                      id="option2"
+                      name="option2"
+                      label="Option 2"
+                      value={curr.option2}
+                      fullWidth
+                      onChange={(e) => handleChange(e, index)}
+                    />
+                  </Grid>
+                  <Grid item xs={6}>
+                    <TextField
+                      required
+                      id="option3"
+                      name="option3"
+                      label="Option3"
+                      value={curr.option3}
+                      fullWidth
+                      onChange={(e) => handleChange(e, index)}
+                    />
+                  </Grid>
+                  <Grid item xs={6}>
+                    <TextField
+                      required
+                      id="option4"
+                      name="option4"
+                      label="Option4"
+                      value={curr.option4}
+                      fullWidth
+                      onChange={(e) => handleChange(e, index)}
+                    />
+                  </Grid>
 
-                    <Grid item xs={6}>
-                      <TextField
-                        id="answer"
-                        name="answer"
-                        label="Answer"
-                        value={curr.answer}
-                        fullWidth
-                        onChange={(e) => handleChange(e, index)}
-                      />
-                    </Grid>
-                    <Grid item xs={3}>
-                      <TextField
-                        id="type"
-                        name="type"
-                        label="Type"
-                        type="number"
-                        value={curr.type}
-                        fullWidth
-                        onChange={(e) => handleChange(e, index)}
-                      />
-                    </Grid>
-                    <Grid item xs={3} style={{ marginTop: 16 }}>
-                      <Button
-                        type="button"
-                        variant="contained"
-                        color="secondary"
-                        onClick={() => RemoveQuestion(index)}
-                      >
-                        Delete
+                  <Grid item xs={6}>
+                    <TextField
+                      id="answer"
+                      name="answer"
+                      label="Answer"
+                      value={curr.answer}
+                      fullWidth
+                      onChange={(e) => handleChange(e, index)}
+                    />
+                  </Grid>
+                  <Grid item xs={3}>
+                    <TextField
+                      id="type"
+                      name="type"
+                      label="Type"
+                      type="number"
+                      value={curr.type}
+                      fullWidth
+                      onChange={(e) => handleChange(e, index)}
+                    />
+                  </Grid>
+                  <Grid item xs={3} style={{ marginTop: 16 }}>
+                    <Button
+                      type="button"
+                      variant="contained"
+                      color="secondary"
+                      onClick={() => RemoveQuestion(index)}
+                    >
+                      Delete
                       </Button>
-                    </Grid>
-                  </React.Fragment>
-                );
-              })
+                  </Grid>
+                </React.Fragment>
+              );
+            })
             : null}
           <Grid item xs={12} style={{ marginTop: 16 }}>
             <Button
