+++ client/src/components/tests/Testform.js
@@ -65,10 +65,10 @@ const TestForm = ({ setTest, handleNext, test }) => {
   };
   return (
     <div style={{ margin: 'auto', marginTop: '1%', maxWidth: 600 }}>
-      <Typography variant="h4" align="center" component="h1" gutterBottom>
+      <Typography style={{ color: "white" }} variant="h4" align="center" component="h1" gutterBottom>
         Create Test
       </Typography>
-      <Typography variant="h5" align="center" component="h2" gutterBottom>
+      <Typography style={{ color: "white" }} variant="h5" align="center" component="h2" gutterBottom>
         Test Form
       </Typography>
